Refactor macros into dedicated modules
