+++ crates/rstest-bdd-macros/src/codegen/mod.rs
@@ -1,0 +1,4 @@
+//! Code generation utilities for the proc macros.
+
+pub(crate) mod scenario;
+pub(crate) mod wrapper;
