+++ crates/rstest-bdd-macros/src/macros/given.rs
@@ -1,0 +1,8 @@
+//! Implementation of the `#[given]` macro.
+
+use proc_macro::TokenStream;
+
+/// Macro for defining a Given step that registers with the step inventory.
+pub(crate) fn given(attr: TokenStream, item: TokenStream) -> TokenStream {
+    super::step_attr(attr, item, rstest_bdd::StepKeyword::Given)
+}
