+++ crates/rstest-bdd-macros/src/parsing/mod.rs
@@ -1,0 +1,4 @@
+//! Utilities for parsing feature files and examples.
+
+pub(crate) mod examples;
+pub(crate) mod feature;
