+++ crates/rstest-bdd-macros/src/utils/mod.rs
@@ -1,0 +1,4 @@
+//! Shared utilities for macro implementation.
+
+pub(crate) mod errors;
+pub(crate) mod fixtures;
