+++ crates/rstest-bdd-macros/src/validation/mod.rs
@@ -1,0 +1,4 @@
+//! Validation helpers for parsing macros.
+
+pub(crate) mod examples;
+pub(crate) mod parameters;
